refactor: make files protected
