+++ autora/cycle/__init__.py
@@ -1,3 +1,4 @@
+from ._simple import SimpleCycle as Cycle
 from .plot_utils import (
     cycle_default_score,
     cycle_specified_score,
@@ -5,4 +6,3 @@ from .plot_utils import (
     plot_results_panel_2d,
     plot_results_panel_3d,
 )
-from .simple import SimpleCycle as Cycle
