+++ autora/cycle/_params.py
@@ -5,8 +5,9 @@ from typing import Dict, Mapping
 
 import numpy as np
 
-from autora.cycle.state import CycleState
 from autora.utils.dictionary import LazyDict
+
+from ._state import CycleState
 
 
 def _resolve_state_params(state: CycleState) -> Dict:
